+++ csrc/ex_c_stuff.h
@@ -4,5 +4,8 @@
 #include "pf_all.h"
 
 char* toCstr(char* ptr, cell_t len);
+void initExFT(void);
+void addLib(char* libName);
+void hello_world(void);
 
 #endif